requirements.txt
Django==3.1.13
djangorestframework==3.12.2
Pillow>=8.2.0

# permissions
django-guardian

# for mac address field
netaddr==0.8.0

# db client for Postgres
psycopg2-binary==2.8.6

git+https://github.com/nerosketch/easysnmp.git#egg=easysnmp
# pid==3.0.4

transliterate==1.10.2

# xml for pay system allpay
djangorestframework-xml==2.0.0

# django-bitfield
git+https://github.com/disqus/django-bitfield.git#egg=django-bitfield

# for clean unused media
django-cleanup==5.1.0

# viberbot
git+https://github.com/Viber/viber-bot-python.git#egg=viberbot

uwsgi_tasks==0.7.3
# redis==3.5.3

django-cors-headers==3.7.0
django-encrypted-model-fields
django-netfields==1.2.2

# pexpect
git+https://github.com/pexpect/pexpect.git#egg=pexpect

django-filter==2.4.0
#pylibmc==1.6.1
djangorestframework-queryfields==1.0.0

djangorestframework-guardian

django-webpush==0.3.3

# linters, and pre commit hook
#flake8==3.9.0
#pre-commit==2.11.1
#autoflake==1.4
#black==20.8b1
#pyupgrade==2.11.0

pyTelegramBotAPI

# radiusapp
git+https://github.com/pyradius/pyrad.git#egg=pyrad

drf-orjson-renderer==1.1.5

pyyaml
